Add files via upload
This HTML file is the front-end interface of the Crop Yield Prediction System, built using Flask. It allows users to input key agricultural data, such as the state, district, crop type, area, season, and year, to predict crop yield and production in kilograms.

This CSS file styles the various elements of the Crop Yield Prediction System's HTML interface, making it visually appealing and easy to navigate. It is located in the static/css directory.